server/pyproject.toml
[project]
name = "statsign-server"
version = "0.1.0"
description = "Local web + BLE controller for Statsign"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "jinja2",
  "pillow",
  "bleak",
]
